Display local PCA

draw.pca displays both local and global scores.

draw.pca(geno, V, sigma, uglob, beg, end, pop, i = 1, j = 2, ancstrl1,
  ancstrl2, adm)

Arguments

geno

a genotype matrix.

V

a loading matrix.

sigma

a vector of singular values.

uglob

a matrix of global scores.

beg

an integer specifying the first marker to be included.

end

an integer specifying the first marker to be excluded.

pop

a list of integers.

i

an integer indicating onto which principal component the individuals are projected when the "scores" option is chosen. Default value is set to 1.

j

an integer indicating onto which principal component the individuals are projected when the "scores" option is chosen. Default value is set to 2.

ancstrl1

an integer.

ancstrl2

an integer.

adm

an integer.

Value

The returned value is a list containing the test statistics and the associated p-values.