pcadapt visualization tool

plot.pcadapt is a method designed for objects of class pcadapt. It provides a plotting utile for quick visualization of pcadapt objects. Different options are currently available : "screeplot", "scores", "stat.distribution", "manhattan" and "qqplot". "screeplot" shows the decay of the genotype matrix singular values and provides a figure to help with the choice of K. "scores" plots the projection of the individuals onto the first two principal components. "stat.distribution" displays the histogram of the selected test statistics, as well as the estimated distribution for the neutral SNPs. "manhattan" draws the Manhattan plot of the p-values associated with the statistic of interest. "qqplot" draws a Q-Q plot of the p-values associated with the statistic of interest.

# S3 method for pcadapt
plot(x, ..., option = "manhattan", K = NULL, i = 1,
  j = 2, pop, gg.col, snp.info = NULL, chr.info = NULL, chr = 1,
  threshold = NULL, by.step = 10, hline = 3, plt.pkg = "ggplot")

Arguments

x
an object of class "pcadapt" generated with pcadapt.

...
…

option
a character string specifying the figures to be displayed. If NULL (the default), all three plots are printed.

K
an integer specifying the principal component of interest. K has to be specified only when using the loadings option.

i
an integer indicating onto which principal component the individuals are projected when the "scores" option is chosen. Default value is set to 1.

j
an integer indicating onto which principal component the individuals are projected when the "scores" option is chosen. Default value is set to 2.

pop
a list of integers or strings specifying which subpopulation the individuals belong to.

gg.col
a list of colors to be used in the score plot.

snp.info
a list containing the names of all genetic markers present in the input.

chr.info
a list containing the chromosome information for each marker.

chr
an integer specifying the chromosome to display.

threshold
for the "qqplot" option, it displays an additional bar which shows the threshold percent of SNPs with smallest p-values and separates them from SNPs with higher p-values.

by.step
an integer.

hline
a numeric value specifying the number of standard deviations above which the z-scores are considered extreme.

plt.pkg
a character string specifying the package to be used to display the graphical outputs. Use "plotly" for interactive plots, or "ggplot" for static plots.

Examples

## see ?pcadapt for examples