File Converter

read.pcadapt converts genotype matrices or files to an appropriate format readable by pcadapt. You may find the converted file in the current directory.

read.pcadapt(input, type, local.env, ploidy, pop.sizes = NULL,
  allele.sep = c("/", "|"), blocksize)

Arguments

input

a genotype matrix or a character string specifying the name of the file to be converted.

type

a character string specifying the type of data to be converted to the pcadapt format. Supported formats are: ped, vcf, lfmm.

local.env

deprecated argument.

ploidy

an integer specifying the ploidy of the individuals.

pop.sizes

a vector specifying the number of individuals to be sampled for each pool.

allele.sep

a vector of characters indicating what delimiters are used in VCF files. By default, only "|" and "/" are recognized.

blocksize

deprecated argument.