Introgression

scan.intro computes statistics to detect excesses of local ancestry based on a PCA approach.

scan.intro(input, K = 2, pop, ancstrl.1, ancstrl.2, admxd, min.maf = 0.05,
  ploidy = 2, window.size = 1000, impute = FALSE)

Arguments

input

a genotype matrix or a character string specifying the name of the file to be imputed.

K

a vector of integers specifying the components along which local ancestries may vary.

pop

a vector of integers or strings specifying which subpopulation the individuals belong to.

ancstrl.1

a string specifying the label of the ancestral population genetically closer to the hybrid population.

ancstrl.2

a string specifying the label of the ancestral population genetically further from the hybrid population.

admxd

a string specifying the label of the hybrid population.

min.maf

a value between 0 and 0.45 specifying the threshold of minor allele frequencies above which p-values are computed.

ploidy

an integer specifying the ploidy of the individuals.

window.size

an integer specifying the window size.

impute

a logical value indicating whether the input data has to imputed.

Value

The returned value is a list containing the test statistics and the associated p-values.